+++ src/main/java/com/example/eternal/repository/manager/ImageRepository.java
@@ -1,6 +1,6 @@
-package com.example.eternal.repository;
+package com.example.eternal.repository.manager;
 
-import com.example.eternal.entity.Image;
+import com.example.eternal.entity.manager.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
