+++ src/main/java/com/example/eternal/repository/manager/NoticeRepository.java
@@ -1,6 +1,6 @@
-package com.example.eternal.repository;
+package com.example.eternal.repository.manager;
 
-import com.example.eternal.entity.Notice;
+import com.example.eternal.entity.manager.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
